fix: adding documentation to the project
adding readme file and fixing some storybooks
