+++ .babelrc
@@ -1,11 +1,6 @@
 {
     "presets": [
-        ["@babel/preset-env",
-        {
-            "targets": {
-                "esmodules": true
-              }
-        }],
+        "@babel/preset-env",
         "@babel/preset-react"
     ]
 }
