+++ package.json
@@ -3,12 +3,12 @@
   "version": "1.0.3",
   "description": "world cup exercise",
   "main": "index.js",
+  "homepage": "https://github.com/ventoji/world-cup-dashboard",
   "scripts": {
-    "build": "webpack --config webpack.config.js",
+    "build:lib": "webpack --config webpack.config.js",
     "build:dev": "webpack --config webpack.dev.js",
     "dev-server": "webpack serve --config webpack.dev.js",
     "storybook": "start-storybook -p 6006",
-    "test": "jest --watchAll",
     "prepublish": "npm run build:lib",
     "ncu": "ncu",
     "ncu:update": "ncu -u",
@@ -16,7 +16,10 @@
     "lint:fix": "eslint src/**/*.{js,jsx} --fix",
     "prettier": "prettier --write src/**/*.{js,jsx,json,css,scss,md}",
     "release": "standard-version",
-    "deploy-storybook": "storybook-to-ghpages",
+    "predeploy": "npm run build-storybook",    
+    "deploy-storybook": "gh-pages -d storybook-static",    
+    "build-storybook": "build-storybook",
+    "test": "jest --watchAll",
     "test:coverage": "npm test -- --coverage --collectCoverageFrom",
     "test:badges": "npm run test:coverage  && jest-coverage-badges"
   },
