--- src/InputWC/InputWC.css
@@ -1,11 +1,0 @@
-.input-wc {
-  margin-bottom: '16px';
-}
-
-.input-wc__error {
-  color: red;
-}
-
-.input-wc__borders {
-  border: 'solid 1px red';
-}
