+++ src/StartGameWC/StartGameWC.js
@@ -1,14 +1,17 @@
 import React, { useState } from 'react';
 import InputWC from '../InputWC';
 import ButtonWC from '../ButtonWC';
+import PropTypes from 'prop-types';
+
+const INITIAL_GAME = {
+  homeTeam: '',
+  awayTeam: '',
+  homeScore: 0,
+  awayScore: 0,
+}
 
 const StartGameWC = ({ onSubmit }) => {
-  const [matchDetails, setMatchDetails] = useState({
-    homeTeam: '',
-    awayTeam: '',
-    homeScore: 0,
-    awayScore: 0,
-  });
+  const [matchDetails, setMatchDetails] = useState(INITIAL_GAME );
 
   const handleMatchDetails = (event) => {
     setMatchDetails({
@@ -42,6 +45,11 @@ const StartGameWC = ({ onSubmit }) => {
       <ButtonWC label="Start new game" onClick={sendMatchDetails} />
     </>
   );
+};
+
+StartGameWC.propTypes = {
+    /** Function to submit new players */
+    onSubmit: PropTypes.func.isRequired
 };
 
 export default StartGameWC;
